Using Get method for family client
As P2N do not gather families correctly since last summer. It seems that it is due to the use of post method in client request (in oposite of the EPO documentation). So this is a quick and dirty hack to make P2N working.
Comments are welcome and needed